requirements.txt
# Core LangChain
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.20
langchain-core==0.1.23

# FastAPI & Web
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1

# Database
sqlalchemy==2.0.25
alembic==1.13.1
aiosqlite==0.19.0

# Vector Store & Search
faiss-cpu==1.7.4
rank-bm25==0.2.2

# Document Processing
pypdf==4.0.1
beautifulsoup4==4.12.3
lxml==5.1.0
unstructured[pdf,html]==0.12.4

# Validation & Config
pydantic==2.5.3
pydantic-settings==2.1.0

# Utilities
python-dotenv==1.0.1
httpx==0.25.2  # Downgraded to be compatible with ollama
tenacity==8.2.3
tiktoken==0.5.2

# Ollama
ollama==0.1.6

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.1
ruff==0.1.14